requirements.txt
anyascii==0.3.0
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.11
django-allauth==0.47.0
django-filter==21.1
django-modelcluster==5.2
django-recaptcha==2.0.6
django-taggit==1.5.1
django-toolbelt==0.0.1
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
l18n==2021.3
oauthlib==3.1.1
openpyxl==3.0.9
Pillow==9.0.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
static3==0.7.0
tablib==3.1.0
telepath==0.2
urllib3==1.26.8
wagtail==2.15.2
wagtail-django-recaptcha==1.0
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
